add nameable accessors and put each decorator in its proper file
