+++ lib/decorator.rb
@@ -1,6 +1,8 @@
 require_relative 'nameable'
 
 class Decorator < Nameable
+  attr_accessor :nameable
+
   def initialize(nameable)
     super()
     @nameable = nameable
@@ -8,17 +10,5 @@ class Decorator < Nameable
 
   def correct_name
     @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    super.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    super[0, 10]
   end
 end
